# stacktrace_libunwind: Use unw_backtrace to compute backtrace faster