Previously, stacktrace_libunwind used unw cursors to compute the backtrace, which is slow. This PR updates stacktrace_libunwind's `GetStackTrace()` function to compute the backtrace using `unw_backtrace`, which is substantially faster.

The main benefit of this change is that malloc profiling becomes substantially faster. I have seen > 5x profiling performance improvements with this change.

I was also able to reproduce @acmorrow's benchmark results from #1277 on my machine. The current libunwind's `GetStackTrace()` implementation performance is as follows:

```
--------------------------------------------------------------------------------
Benchmark                      Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------
BM_GetStackTrace/1          3832 ns         3828 ns       182886 items_per_second=261.232k/s
BM_GetStackTrace/8          7705 ns         7692 ns        95400 items_per_second=130.008k/s
BM_GetStackTrace/64        31017 ns        31003 ns        22564 items_per_second=32.2546k/s
BM_GetStackTrace/512      184911 ns       184846 ns         3789 items_per_second=5.4099k/s
BM_GetStackTrace/4096    1356384 ns      1353508 ns          491 items_per_second=738.821/s
BM_GetStackTrace/8192    2578422 ns      2564196 ns          273 items_per_second=389.986/s
```

With this PR, the performance improves substantially:

```
--------------------------------------------------------------------------------
Benchmark                      Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------
BM_GetStackTrace/1           102 ns          102 ns      6913378 items_per_second=9.84186M/s
BM_GetStackTrace/8           166 ns          166 ns      4218120 items_per_second=6.02384M/s
BM_GetStackTrace/64          517 ns          517 ns      1355825 items_per_second=1.93412M/s
BM_GetStackTrace/512        3252 ns         3251 ns       216073 items_per_second=307.559k/s
BM_GetStackTrace/4096      25256 ns        25194 ns        27830 items_per_second=39.6917k/s
BM_GetStackTrace/8192      50328 ns        50238 ns        13935 items_per_second=19.9051k/s
```

I have tested this change with `make check` and `make stacktrace_unittest && ./stacktrace_unittest`.

I would greatly appreciate any feedback regarding this PR. Thank you!

Closes #1277