@alk 2024-08-06
Thanks a lot for the patch.

But the code appears to be still doing unbounded caching within libunwind. In practice it won't grow much, usually, but who knows. I.e.: https://github.com/libunwind/libunwind/blob/fb917e18e8bef992ddd4179143a65a59f18f00ca/src/x86_64/Gtrace.c#L274

So perhaps we make it optional? Also I see unw_backtrace2 which can unwind from unw_context.

As for the patch; the max backtrace length thingy and fprintf with exit() is not great. We have places where we do quite a bit of backtracing depth, more than 32. Also you sure you want to do a static per-thread temporary backtraces buffer?