@thecodingwizard 2024-08-07
Thank you for taking a look!

> So perhaps we make it optional?

What's the best way to do this? Maybe make a `stacktrace_libunwind-fast-inl.h` file and add `libunwind_fast` as a new stacktrace option? (For the other three functions that aren't `GetStackTrace()`, we can call the libunwind versions of those functions to avoid code duplication.)

> Also I see unw_backtrace2 which can unwind from unw_context.

Perhaps I'll add a comment for now and leave that for a second PR? Please let me know if you'd rather I do both at once though.

> As for the patch; the max backtrace length thingy and fprintf with exit() is not great.

I agree. Do you know of a better way to do this? I wasn't sure whether using `malloc()` inside this function would be safe (and even if it were, I was worried that would be too slow). However, I'm not sure how to properly handle `skip_count` without using a temporary array.

Thanks again for your feedback!