@alk 2024-08-07
Yes, splitting this into new distinct stacktracing method is fine.

As for how to deal with the storage, how about doing it on stack. Say 256 or 128 entries. Then whatever doesn't fit, gets truncated (no big deal). Another easy win here is if you're given more storage than whatever you have on stack, use it rather than your storage.