@alk 2024-09-15
I apologize for loosing track of this.

So looking at it just now I have few comments. But main one is libunwind by default already aliases backtrace to unw_backtrace. So then why not use some of the tricks you did here to "unbreak" stacktrace_generic instead ? (It was for long time hardcoded but since 2019 libunwind has configure flag to disable it: https://github.com/libunwind/libunwind/pull/142)